#6 Save to kintone
